requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
jinja2==3.1.2

# Database Drivers
motor==3.3.2
pymongo==4.6.1
pinecone-client==3.0.2

# LLM and AI
groq==0.4.2
langgraph==0.0.34
langchain-core==0.1.23
langchain-groq==0.0.1
sentence-transformers==2.3.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment & Config
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
numpy==1.26.3
